demo

+++ internal/cluster/kubernetes.go
@@ -217,11 +217,21 @@ func (c *NativeKubernetesCluster) DownCluster(ctx context.Context) error {
 	// If more thorough cleanup is needed, additional logic can be added here.
 
 	ns := generateInstallNs(c.cluster)
-	err = ctrlClient.Delete(ctx, ns)
-	if err != nil && !apierrors.IsNotFound(err) {
+	err = ctrlClient.Get(ctx, client.ObjectKey{Name: ns.Name}, ns)
+	if err != nil {
 		if apierrors.IsNotFound(err) {
 			return nil
 		}
+		return errors.Wrap(err, "failed to get namespace")
+	}
+
+	if ns.DeletionTimestamp != nil {
+		// Namespace is already being deleted
+		return errors.New("waiting for namespace deletion")
+	}
+
+	err = ctrlClient.Delete(ctx, ns)
+	if err != nil {
 		return errors.Wrap(err, "failed to delete namespace")
 	}
